Added some more permission configurability for the witchdoctor. The UI can now be opened from console using /wd open [player]. Also made the exising /bv give command work from console as long as there is a target player.

+++ src/main/java/io/gitlab/arkdirfe/boxedvillagers/ui/WitchdoctorGuiManager.java
@@ -139,11 +139,15 @@ public class WitchdoctorGuiManager implements Listener
         if(slotMovable && cursorEmpty && event.isShiftClick() && event.isLeftClick()) // Purge trade
         {
             event.setCancelled(true);
-            controller.purgeTrade(slotIndex);
-            view.setCursor(new ItemStack(Material.AIR));
+            
+            if(controller.hasPurgePerms())
+            {
+                controller.purgeTrade(slotIndex);
+                view.setCursor(new ItemStack(Material.AIR));
+            }
         }
         
-        if(slotMovable && cursorEmpty && event.isShiftClick() && event.isRightClick())
+        if(slotMovable && cursorEmpty && event.isShiftClick() && event.isRightClick()) // Extract trade
         {
             event.setCancelled(true);
